check & catch exception if network has not been available

+++ app/src/main/java/com/vn/hcmute/team/cortana/mymoney/data/remote/RemoteRepository.java
@@ -35,6 +35,9 @@ public class RemoteRepository implements RemoteTask.UserTask, RemoteTask.ImageTa
     @Override
     public Observable<User> login(UserCredential userCredential) {
         UserService userService = mServiceGenerator.getService(UserService.class);
+        if(userService ==null){
+            return null;
+        }
         return userService.login(userCredential)
                   .map(new Function<JsonResponse<User>, User>() {
                       
@@ -53,6 +56,9 @@ public class RemoteRepository implements RemoteTask.UserTask, RemoteTask.ImageTa
     @Override
     public Observable<String> register(User user) {
         UserService userService = mServiceGenerator.getService(UserService.class);
+        if(userService ==null){
+            return null;
+        }
         return userService.register(user)
                   .map(new Function<JsonResponse<String>, String>() {
                       @Override
@@ -71,6 +77,9 @@ public class RemoteRepository implements RemoteTask.UserTask, RemoteTask.ImageTa
     @Override
     public Observable<List<Image>> getImage(String userid, String token) {
         ImageService imageService = mServiceGenerator.getService(ImageService.class);
+        if(imageService ==null){
+            return null;
+        }
         return imageService.get(userid, token)
                   .map(new Function<JsonResponse<List<Image>>, List<Image>>() {
                       @Override
@@ -89,7 +98,9 @@ public class RemoteRepository implements RemoteTask.UserTask, RemoteTask.ImageTa
     @Override
     public Observable<Image> getImageByid(String userid, String token, String imageid) {
         ImageService imageService = mServiceGenerator.getService(ImageService.class);
-        
+        if(imageService ==null){
+            return null;
+        }
         return imageService.getImageById(userid, token, imageid)
                   .map(new Function<JsonResponse<Image>, Image>() {
                       @Override
@@ -108,7 +119,9 @@ public class RemoteRepository implements RemoteTask.UserTask, RemoteTask.ImageTa
     @Override
     public Observable<String> removeImage(String userid, String token, String imageid) {
         ImageService imageService = mServiceGenerator.getService(ImageService.class);
-        
+        if(imageService ==null){
+            return null;
+        }
         return imageService.remove(userid, token, imageid)
                   .map(new Function<JsonResponse<String>, String>() {
                       @Override
@@ -127,6 +140,9 @@ public class RemoteRepository implements RemoteTask.UserTask, RemoteTask.ImageTa
     public Observable<String> uploadImage(RequestBody userid, RequestBody token, RequestBody detail,
               MultipartBody.Part file) {
         ImageService imageService = mServiceGenerator.getService(ImageService.class);
+        if(imageService ==null){
+            return null;
+        }
         return imageService.upload(userid, token, detail, file)
                   .map(new Function<JsonResponse<String>, String>() {
                       @Override
@@ -144,7 +160,9 @@ public class RemoteRepository implements RemoteTask.UserTask, RemoteTask.ImageTa
     @Override
     public Observable<String> updateImage(String userid, String token, String imageid) {
         ImageService imageService = mServiceGenerator.getService(ImageService.class);
-        
+        if(imageService ==null){
+            return null;
+        }
         return imageService.update(userid, token, imageid)
                   .map(new Function<JsonResponse<String>, String>() {
                       @Override
